refactor(avatar): extract AvatarWarningModal component
